[Enhance] Logger: set, store and restore level

+++ utils/file.py
@@ -1,6 +1,7 @@
 import shutil
 
 
-def rmtree_and_mkdir(path):
-    shutil.rmtree(path, ignore_errors=True)
+def rmtree_and_mkdir(path, overwrite=False):
+    if overwrite:
+        shutil.rmtree(path, ignore_errors=True)
     path.mkdir(parents=True, exist_ok=True)
